PyLL has been made timezone-aware by pytz.
That way users can safely assume all their dates and expiries are
processed in a defined timezone - most probably their's - no matter
what timezone the server is configured to.

Quickstart asks for the user's timezone.

Furthermore, de-jure "now" is the moment a command has been issued.
This is because the user cannot predict (in this case) how long a
process will take.